seperate field and field message
